LayoutTests/fast/masking/parsing-mask-expected.txt

Test that clip-path shapes accept different length units

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS innerStyle("mask", "none") is "none"
PASS innerStyle("mask", "none, none") is "none, none"
PASS innerStyle("mask", "none, none, none") is "none, none, none"
PASS innerStyle("mask", "url(file:///image.png), none") is "url(\"file:///image.png\"), none"
PASS innerStyle("mask", "none, url(file:///image.png)") is "none, url(\"file:///image.png\")"
PASS innerStyle("mask", "top left") is "left top"
PASS innerStyle("mask", "bottom right") is "right bottom"
PASS innerStyle("mask", "left bottom") is "left bottom"
PASS innerStyle("mask", "right top") is "right top"
PASS innerStyle("mask", "center") is "center center"
PASS innerStyle("mask", "none top") is "center top"
PASS innerStyle("mask", "none bottom") is "center bottom"
PASS innerStyle("mask", "none right") is "right center"
PASS innerStyle("mask", "none top right") is "right top"
PASS innerStyle("mask", "none bottom left") is "left bottom"
PASS innerStyle("mask", "none right") is "right center"
PASS innerStyle("mask", "none left") is "left center"
PASS innerStyle("mask", "center 50%") is "center 50%"
PASS innerStyle("mask", "50px 50%") is "50px 50%"
PASS innerStyle("mask", "center left") is "left center"
PASS innerStyle("mask", "top center") is "center top"
PASS innerStyle("mask", "left 10px top 15px") is "left 10px top 15px"
PASS innerStyle("mask", "left 10% top 30%") is "left 10% top 30%"
PASS innerStyle("mask", "right top 15px") is "right top 15px"
PASS innerStyle("mask", "left 10px center") is "left 10px center"
PASS innerStyle("mask", "center top 20px") is "center top 20px"
PASS innerStyle("mask", "center left 30px") is "left 30px center"
PASS innerStyle("mask", "left 20% top") is "left 20% top"
PASS innerStyle("mask", "center center") is "center center"
PASS innerStyle("mask-position", "left 10px top 15px") is "left 10px top 15px"
PASS innerStyle("mask-position", "left 10% top 30%") is "left 10% top 30%"
PASS innerStyle("mask-position", "right top 15px") is ""
PASS innerStyle("mask-position", "left 10px center") is ""
PASS innerStyle("mask-position", "center top 20px") is ""
PASS innerStyle("mask-position", "center left 30px") is ""
PASS innerStyle("mask-position", "left 20% top") is ""
PASS innerStyle("mask", "none alpha") is "alpha"
PASS innerStyle("mask", "none luminance, none alpha") is "luminance, alpha"
PASS innerStyle("mask", "none match-source, none, none luminance") is "none, none, luminance"
PASS innerStyle("mask-mode", "alpha") is "alpha"
PASS innerStyle("mask-mode", "luminance") is "luminance"
PASS innerStyle("mask-mode", "match-source") is "match-source"
PASS computedStyle("mask-mode", "alpha") is "alpha"
PASS computedStyle("mask-mode", "match-source") is "match-source"
PASS computedStyle("mask-mode", "luminance") is "luminance"
PASS computedStyle("mask-mode", "") is "match-source"
PASS innerStyle("mask", "repeat-x") is "repeat-x"
PASS innerStyle("mask", "repeat-y") is "repeat-y"
PASS innerStyle("mask", "repeat") is "repeat"
PASS innerStyle("mask", "space") is "space"
PASS innerStyle("mask", "no-repeat") is "no-repeat"
PASS innerStyle("mask", "repeat space") is "repeat space"
PASS innerStyle("mask", "repeat round") is "repeat round"
PASS innerStyle("mask", "repeat no-repeat") is "repeat-x"
PASS innerStyle("mask", "repeat space, repeat-x") is "repeat space, repeat-x"
PASS innerStyle("mask", "repeat none") is "repeat"
PASS innerStyle("mask", "none repeat") is "repeat"
PASS innerStyle("mask", "padding-box") is "padding-box"
PASS innerStyle("mask", "border-box") is "none"
PASS innerStyle("mask", "content-box") is "content-box"
PASS innerStyle("mask", "padding-box none") is "padding-box"
PASS innerStyle("mask", "none padding-box") is "padding-box"
PASS innerStyle("mask", "padding-box content-box") is "padding-box content-box"
PASS innerStyle("mask", "content-box content-box") is "content-box"
PASS innerStyle("mask", "padding-box border-box") is "padding-box border-box"
PASS innerStyle("mask", "padding-box border-box none") is "padding-box border-box"
PASS innerStyle("mask", "none padding-box border-box") is "padding-box border-box"
PASS innerStyle("mask", "none left top / auto") is "left top"
PASS innerStyle("mask", "none left top / auto auto") is "left top"
PASS innerStyle("mask", "none left top / 100%") is "left top / 100%"
PASS innerStyle("mask", "none left top / 100% 100%") is "left top / 100% 100%"
PASS innerStyle("mask", "none left top / 0%") is "left top / 0%"
PASS innerStyle("mask", "none left top / auto 0%") is "left top / auto 0%"
PASS innerStyle("mask", "none left top / cover") is "left top / cover"
PASS innerStyle("mask", "none left top / contain") is "left top / contain"
PASS innerStyle("mask", "none left 20px top 10px / contain") is "left 20px top 10px / contain"
PASS innerStyle("mask", "none left 20px top / contain") is "left 20px top / contain"
PASS innerStyle("mask", "none padding-box content-box") is "padding-box content-box"
PASS innerStyle("mask", "none padding-box") is "padding-box"
PASS innerStyle("mask", "none top") is "center top"
PASS innerStyle("mask", "none center right 20px") is "right 20px center"
PASS innerStyle("mask", "none border-box left top") is "left top"
PASS innerStyle("mask", "none border-box left top 20px") is "left top 20px"
PASS innerStyle("mask", "none border-box content-box left top repeat-x") is "left top repeat-x border-box content-box"
PASS innerStyle("mask", "none border-box content-box left top / auto repeat-x") is "left top repeat-x border-box content-box"
PASS innerStyle("mask", "none border-box content-box right 0px center / auto repeat-x") is "right 0px center repeat-x border-box content-box"
PASS computedStyle("mask", "top left") is "none"
PASS computedStyle("mask", "bottom right") is "none"
PASS computedStyle("mask", "left bottom") is "none"
PASS computedStyle("mask", "right top") is "none"
PASS computedStyle("mask", "center") is "none"
PASS computedStyle("mask", "none top") is "none"
PASS computedStyle("mask", "none bottom") is "none"
PASS computedStyle("mask", "none right") is "none"
PASS computedStyle("mask", "none top right") is "none"
PASS computedStyle("mask", "none bottom left") is "none"
PASS computedStyle("mask", "none right") is "none"
PASS computedStyle("mask", "none left") is "none"
PASS computedStyle("mask", "center 50%") is "none"
PASS computedStyle("mask", "50px 50%") is "none"
PASS computedStyle("mask", "center left") is "none"
PASS computedStyle("mask", "top center") is "none"
PASS computedStyle("mask", "left 10px top 15px") is "none"
PASS computedStyle("mask", "left 10% top 30%") is "none"
PASS computedStyle("mask", "right top 15px") is "none"
PASS computedStyle("mask", "left 10px center") is "none"
PASS computedStyle("mask", "center top 20px") is "none"
PASS computedStyle("mask", "center left 30px") is "none"
PASS computedStyle("mask", "left 20% top") is "none"
PASS computedStyle("mask", "center center") is "none"
PASS computedStyle("mask", "repeat-x") is "none"
PASS computedStyle("mask", "repeat-y") is "none"
PASS computedStyle("mask", "repeat") is "none"
PASS computedStyle("mask", "space") is "none"
PASS computedStyle("mask", "no-repeat") is "none"
PASS computedStyle("mask", "repeat space") is "none"
PASS computedStyle("mask", "repeat round") is "none"
PASS computedStyle("mask", "repeat no-repeat") is "none"
PASS computedStyle("mask", "repeat none") is "none"
PASS computedStyle("mask", "none repeat") is "none"
PASS computedStyle("mask", "padding-box") is "none"
PASS computedStyle("mask", "border-box") is "none"
PASS computedStyle("mask", "content-box") is "none"
PASS computedStyle("mask", "padding-box none") is "none"
PASS computedStyle("mask", "none padding-box") is "none"
PASS computedStyle("mask", "padding-box content-box") is "none"
PASS computedStyle("mask", "content-box content-box") is "none"
PASS computedStyle("mask", "padding-box border-box") is "none"
PASS computedStyle("mask", "padding-box border-box none") is "none"
PASS computedStyle("mask", "none padding-box border-box") is "none"
PASS computedStyle("mask", "none left top / auto") is "none"
PASS computedStyle("mask", "none left top / auto auto") is "none"
PASS computedStyle("mask", "none left top / 100%") is "none"
PASS computedStyle("mask", "none left top / 0%") is "none"
PASS computedStyle("mask", "none left top / auto 0%") is "none"
PASS computedStyle("mask", "none left top / cover") is "none"
PASS computedStyle("mask", "none left top / contain") is "none"
PASS computedStyle("mask", "none left 20px top 10px / contain") is "none"
PASS computedStyle("mask", "none left 20px top / contain") is "none"
PASS computedStyle("mask", "none padding-box content-box") is "none"
PASS computedStyle("mask", "none padding-box") is "none"
PASS computedStyle("mask", "none top") is "none"
PASS computedStyle("mask", "none center right 20px") is "none"
PASS computedStyle("mask", "none border-box left top") is "none"
PASS computedStyle("mask", "none border-box left top 20px") is "none"
PASS computedStyle("mask", "none border-box content-box left top repeat-x") is "none"
PASS computedStyle("mask", "none border-box content-box left top / auto repeat-x") is "none"
PASS computedStyle("mask", "none border-box content-box right 0px center / auto repeat-x") is "none"
PASS innerStyle("mask", "top none left") is ""
PASS innerStyle("mask", "right none bottom") is ""
PASS innerStyle("mask", "right right") is ""
PASS innerStyle("mask", "left left") is ""
PASS innerStyle("mask", "top top") is ""
PASS innerStyle("mask", "bottom bottom") is ""
PASS innerStyle("mask", "50% none 50%") is ""
PASS innerStyle("mask", "repeat-x repeat-x") is ""
PASS innerStyle("mask", "space repeat-y") is ""
PASS innerStyle("mask", "repeat space space") is ""
PASS innerStyle("mask", "padding-box border-box content-box") is ""
PASS innerStyle("mask", "none / auto") is ""
PASS innerStyle("mask", "none repeat-x / auto") is ""
PASS innerStyle("mask", "none border-box / auto") is ""
PASS innerStyle("mask", "none top left / cover 100%") is ""
PASS innerStyle("mask", "scroll") is ""
PASS innerStyle("mask", "fixed") is ""
PASS innerStyle("mask", "local") is ""
PASS innerStyle("mask", "space scroll") is ""
PASS innerStyle("mask", "none scroll") is ""
PASS innerStyle("mask", "none top left / auto repeat-x scroll border-box border-box") is ""
PASS innerStyle("mask", "right top left") is ""
PASS innerStyle("mask", "center left center") is ""
PASS innerStyle("mask", "center top center") is ""
PASS innerStyle("mask", "center right bottom") is ""
PASS innerStyle("mask", "top solid bottom") is ""
PASS innerStyle("mask", "none top left right center top / auto repeat-x scroll border-box border-box") is ""
PASS innerStyle("mask", "none center center 20px / auto repeat-x scroll border-box border-box") is ""
PASS innerStyle("mask", "none top 20px right 30px center / auto repeat-x scroll border-box border-box") is ""
PASS innerStyle("mask", "none top 20px top 30px / auto repeat-x scroll border-box border-box") is ""
PASS innerStyle("mask", "none top 20px bottom / auto repeat-x scroll border-box border-box") is ""
PASS innerStyle("mask", "none alpha, none match-source alpha") is ""
PASS innerStyle("mask", "none alpha luminance") is ""
PASS innerStyle("mask-mode", "auto") is ""
PASS innerStyle("mask-mode", "rubbish") is ""
PASS innerStyle("mask-mode", "") is ""
PASS innerStyle("mask-mode", "center") is ""
PASS innerStyle("mask-mode", "repeat") is ""
PASS innerStyle("mask-position-x", "left") is ""
PASS innerStyle("mask-position-y", "top") is ""
PASS innerStyle("mask-repeat-x", "repeat") is ""
PASS innerStyle("mask-repeat-y", "repeat") is ""
PASS successfullyParsed is true

TEST COMPLETE
